dohCluster

function for the speaq package. The function dohCluster will be automatically called by the function speaq_align. The speaq package is used to perform the signal alignment across the individual spectra. For speaq please cite: (Beirnaert C, Meysman P, Vu TN, Hermans N, Apers S, Pieters L, et al. (2018) speaq 2.0: A complete workflow for high-throughput 1D NMRspectra processing and quantification. PLoS Comput Biol 14(3): e1006018. https://doi.org/10.1371/journal.pcbi.1006018. The spectra deconvolution process yields the signals of all spectra. Due to slight changes in measurement conditions, e.g. pH variations, signal positions may vary slightly across spectra. As a consequence, prior to further analysis signals belonging to the same compound have to be aligned across spectra. This is the purpose of the speaq package.

Usage

dohCluster(
  X,
  peakList,
  refInd = 0,
  maxShift = 100,
  acceptLostPeak = TRUE,
  verbose = TRUE
)

Arguments

X

data_matrix as returned by gen_feat_mat()

peakList

peakList as returned by function gen_feat_mat()

refInd

(positive integer) the number of the reference spectrum i.e. the spectrum to which all signals will be aligned to. This number will be automatically determined by the function speaq_align which then calls dohCluster

maxShift

(positive integer) maximum number of points along the "ppm-axis" which a value can be moved by speaq package e.g. 50

acceptLostPeak

(logic) default is TRUE

verbose

(logic) default is TRUE

Details

The function dohCluster of the speaq package has been rewritten to be compatible with the data generated by MetaboDecon1D and the function gen_feat_mat and to return a new peakList of aligned spectra. Overwrite original dohCluster function of speaq package. Function is able to return the new peakList of aligned spectra.